requirements.txt
BAC0>=2025.9.15
influxdb>=5.3.2
paho-mqtt>=2.0.0